Don't use impl Context for implementation details
Methods are identified by the type that they're implemented on, not the module that they're implemented in. This means that two methods in separate files can collide even if they're private.

I ran into this when implementing UnicodeKeyboardContext for X11 because I had two separate modules with a key_event function in them. I changed the others for consistency.