Server doesn't need tokio
Since the server is only meant to handle one connection at a time, there's nothing else to do while waiting for messages. This means that it's basically single-threaded and synchronous anyway so we might as well drop tokio. This was pretty painless because the standard library and tokio are nearly identical (apart from async obviously).