Enum overhaul
Having a separate binary that generated code seemed really weird so I'm using a macro instead. The issue is that the same level of automation is not possible. We can't create an identifier from a string. We could use a proc-macro but the issue with proc-macro generated code is that it doesn't work with intellisense. So the result is that we have to have an identifier and a display name string instead of generating one from the other.

The enums implement a trait that provides reflection capabilities. The generator is now generating Swift enums from the Rust enums.