Allow negative numbers to be entered with TextInput
Hijacking the decimal point button like this is a bit sneaky but it's the easiest way to dealing with this. Lucky we don't need to deal with negative floats!

At this point, EditCommandVC is basically done. That's the most complicated part of the new configuration system out of the way.