Enable encryption
What I have right now is basically a POC. The server is ready and the encryption works. I need to handle the UI part of it. This means asking the user to scan a QR code. I'm not completely convinced that a QR code is the best way to do this though.