Add QR code scanner view controller
This is very close to fully working. I just need to figure out some details about connection status and having secure mode enabled but the key not set. It also might be a bit annoying to scan the code every time the app connects to the server. If the server stores aside the key, then where will it store it?